+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.2.0 (March 6, 2020)
+* add setScore method
+* add clearSuspendData method
+* the following methods now return promises
+  * getSuspendData
+  * setSuspendData
+  * clearSuspendData
+  * setStatus
+  * setScore
+  * set
+* add jest / enzyme test suite
+
 # 0.1.6 (Feb 19, 2020)
 * Transfer repository ownership to S4 NetQuest
   * Add related images and links for S4 NetQuest